New subtype - State <: StaticArrays.FieldVector{2, Int}
This is a change that should not affect the behavior of this notebook directly. It makes the notebook's MDP easier to use with other libraries including Crux.jl.

See https://github.com/JuliaPOMDP/QuickPOMDPs.jl/issues/41#issuecomment-2140731919 for some context. In short, this change results in an automatic definition for POMDPs.convert_s, which Crux relies on.